feat: Add feature generation
Turns out that heightmap data is required for rendering a chunk on the client.
Also added caching to the blocking `getChunk` world generation.
And added a mixin into `WorldAccess` to ensure `isChunkLoaded` returns a valid result. Will need to optimise it further by adding a method for checking the cache to `NoisiumServerWorldChunkManager`.