fix: Fix feature generation
Uses a workaround that returns a dummy `ProtoChunk` in `ChunkRegion#getChunk` to prevent issues due to missing `null` checks.
This means there are issues at chunk borders, where features will get cut off. A potential solution for this is to store features' block modifications in the save data, and apply these block modifications on top of world generation chunk data.